+++ core/internal/cltest/factories.go
@@ -406,12 +406,14 @@ func MustAddRandomKeyToKeystore(t testing.TB, ethKeyStore keystore.Eth) (ethkey.
 
 func MustAddKeyToKeystore(t testing.TB, key ethkey.KeyV2, chainID *big.Int, ethKeyStore keystore.Eth) {
 	t.Helper()
-	err := ethKeyStore.Add(key, chainID)
+	ethKeyStore.XXXTestingOnlyAdd(key)
+	err := ethKeyStore.Enable(key, chainID)
 	require.NoError(t, err)
 }
 
 // MustInsertRandomKey inserts a randomly generated (not cryptographically
 // secure) key for testing
+// By default it is enabled for the fixture chain
 func MustInsertRandomKey(
 	t testing.TB,
 	keystore keystore.Eth,
@@ -419,35 +421,36 @@ func MustInsertRandomKey(
 ) (ethkey.KeyV2, common.Address) {
 	t.Helper()
 
-	chainID := *utils.NewBig(&FixtureChainID)
+	chainIDs := []utils.Big{*utils.NewBig(&FixtureChainID)}
 	for _, opt := range opts {
 		switch v := opt.(type) {
 		case utils.Big:
-			chainID = v
+			chainIDs[0] = v
+		case []utils.Big:
+			chainIDs = v
 		}
 	}
 
 	key := MustGenerateRandomKey(t)
-	require.NoError(t, keystore.Add(key, chainID.ToInt()))
-	state, err := keystore.GetState(key.ID())
-	require.NoError(t, err)
+	keystore.XXXTestingOnlyAdd(key)
 
-	for _, opt := range opts {
-		switch v := opt.(type) {
-		case int:
-			state.NextNonce = int64(v)
-		case int64:
-			state.NextNonce = v
-		case bool:
-			state.IsFunding = v
-		case utils.Big:
-			state.EVMChainID = v
-		default:
-			t.Fatalf("unrecognised option type: %T", v)
+	for _, cid := range chainIDs {
+		require.NoError(t, keystore.Enable(key, cid.ToInt()))
+
+		var nonce int64
+		for _, opt := range opts {
+			switch v := opt.(type) {
+			case int:
+				nonce = int64(v)
+			case int64:
+				nonce = v
+			default:
+				t.Fatalf("unrecognised option type: %T", v)
+			}
 		}
+		err := keystore.Reset(key, cid.ToInt(), nonce)
+		require.NoError(t, err)
 	}
-	err = keystore.SetState(state)
-	require.NoError(t, err)
 
 	return key, key.Address.Address()
 }